Store results of param callbacks. This allows checking those params for a value of any kind. Fixes #144
We need a way to be able to check if a param callback has a value whether they passed a static value or used a callback. We now output-buffer the results of those params and store them in an array as a field object property. Unit tests have been added to verify this behavior.