Update CMB2_Types.php
CMB2 has a strange bug that creates an extra "select" dropdown field when the "repeatable" argument is set to true for a "select" field and it has more than 1 values. In simple, if a select field has been repeated N times where N>1, this bug always renders an extra select dropdown at the end. 

This bug is only related to "select" fields, specially when their "repeatable" is set to true. Doesn't affect other fields.