Add validation for radio, select and select_timezone fields.
Validate the received value against the allowed options as provided in the field definition.
Returns false if the received value is invalid.

Uses the `options()` method to retrieve options. This allows for options provided via the `options_cb` parameter.

Includes filter to allow for dynamically altering the options array used for validating radio & select fields.