File-upload field now attaches to current post
Until now, the newly uploaded files were not attached to the current post (if we are in a post setting). This was due to the fact that the media frame fell-back to the 'select' model (the most basic one). When a post ID is available we should use the 'post' model.