Get and use the correct preview image sizes for 'file' and 'file_list' fields
In CMB2_Type_File[_List].php – try to get the closest named image size
from the ‘preview_size’ array and get the correct width and height from
the named size. Pass the named size string as well as the dimensions to
the hidden field to be used by the JS media uploader.

In cmb2.js we can then retrieve the correct image (url and dimensions)
form the returned JS attachment object. Added logic to fit the image to
the preview size rather than stretching the image.

Updated the default preview sizes from the hardcoded 350x350px and
50x50px to ‘large’ and ‘thumbnail’.

Fixes #844