Add function_exists('add_action') check
Add function_exists('add_action') check to initialization file. When
loading CMB2 via Composer (and autoloading init.php), the init.php will
be loaded every time. Usually this is not a problem, however if we are
also using Composer to load a PHP development tool (e.g. PHP Codesniffer
(phpcs)), when the development tool runs it is NOT in the context of
WordPress.

Running CMB2's init.php outside of the context of WordPress
results in a fatal error because the function add_action() does not
exist. By adding this function_exists() check, we can avoid this fatal
error when autoloading init.php outside of the context of WordPress.

See #1270